refactor(src): Verifica cópias de imagens
Ampliar a capacidade de apuração da função verficar_pasta_existente para
imagens também.
